+++ index.js
@@ -173,7 +173,7 @@ function buildHBSTemplate(tree, inputParsedPath, vars, svelteOptions) {
     {{#let (component this.svelteContent) as |SvelteContent|}}
       <SvelteContent ...attributes>
         {{#if this._showStartBound}}<span {{did-insert this.getStartBound}}></span>{{/if}}
-        {{yield}}
+        {{#if this.defaultSlotElement}}{{#in-element this.defaultSlotElement nextSibling=this.defaultSlotAnchor}}{{#if this.showsDefaultSlot}}{{yield}}{{/if}}{{/in-element}}{{/if}}
         {{#if this._showEndBound}}<span {{did-insert this.getEndBound}} {{did-insert this.insertSvelteComponent}}></span>{{/if}}
         {{did-update this.updateSvelteComponent}}
         {{will-destroy this.teardownSvelteComponent}}