+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Modified
+- Replace document fragment hack with direct Glimmer integration for default slot
+
 ## [0.1.1] - 2022-07-01
 ### Modified
 - Fixed regression preventing tagless component rendering